+++ apps/io-services-cms-webapp/src/main.ts
@@ -1,7 +1,13 @@
 import { ServiceLifecycle, stores } from "@io-services-cms/models";
+import * as O from "fp-ts/Option";
+import * as RA from "fp-ts/ReadonlyArray";
+import * as RR from "fp-ts/ReadonlyRecord";
 import { pipe } from "fp-ts/lib/function";
 import { getConfigOrThrow } from "./config";
-import { expressToAzureFunction } from "./lib/azure/adapters";
+import {
+  expressToAzureFunction,
+  toAzureFunctionHandler,
+} from "./lib/azure/adapters";
 import { getDatabase } from "./lib/azure/cosmos";
 import { getApimClient } from "./lib/clients/apim-client";
 import { jiraClient } from "./lib/clients/jira-client";
@@ -9,9 +15,12 @@ import { createRequestReviewHandler } from "./reviewer/request-review-handler";
 import { createReviewCheckerHandler } from "./reviewer/review-checker-handler";
 import { apimProxy } from "./utils/apim-proxy";
 import { jiraProxy } from "./utils/jira-proxy";
-import { getDao } from "./utils/service-review-dao";
 import { createWebServer } from "./webservice";
-import { handler as OnLegacyServiceChangeHandler } from "./watchers/on-legacy-service-change";
+
+import { processBatchOf, setBindings } from "./lib/azure/misc";
+import { handler as onServiceLifecycleChangeHandler } from "./watchers/on-services-lifecycles-change";
+
+import { getDao } from "./utils/service-review-dao";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-unused-vars
 const BASE_PATH = require("../host.json").extensions.http.routePrefix;
@@ -59,4 +68,16 @@ export const serviceReviewCheckerEntryPoint = createReviewCheckerHandler(
   serviceLifecycleStore
 );
 
-export const onLegacyServiceChangeEntryPoint = OnLegacyServiceChangeHandler;
+export const onServiceLifecycleChangeEntryPoint = pipe(
+  onServiceLifecycleChangeHandler,
+  processBatchOf(ServiceLifecycle.ItemType),
+  setBindings((results) => ({
+    requestReview: pipe(
+      results,
+      RA.map(RR.lookup("requestReview")),
+      RA.filter(O.isSome),
+      RA.map(JSON.stringify)
+    ),
+  })),
+  toAzureFunctionHandler
+);