+++ packages/io-services-cms-models/src/index.ts
@@ -1,2 +1,4 @@
+export { FSMStore, stores } from "./lib/fsm";
+export * as Queue from "./queue";
 export * as ServiceLifecycle from "./service-lifecycle";
-export { stores, FSMStore } from "./lib/fsm";
+export * as ServicePublication from "./service-publication";